Make things a lot faster
Pros: was surprised/had forgotten just how many threads are in the default pool for Concurrent Ruby (java's max int value), and it turns out to be way, way faster than just setting it to the max processor count

Also the chunking of tests to simulate workers was wrong and never worked. Oops

Anyway, things are a lot faster this way, which is good, and required me to adjust the --fail-fast test a bit. However, now "--workers" is kind of a misnomer, since most people think of that as processes and this is very much threads. should consult with @tenderlove