Failing test to illustrate a problem with how TLDR gathers tests
TLDR gathers tests by recursively looking for descendants of the TLDR class. If you have multiple Rake tasks that look in different places for tests, that's all well and good, but the first Rake task that runs (because it's all in the same process) will load all of its tests and the second Rake task will, of course, find them, because they're still defined.

Decoupling path from test definition is generally good but in this particular case it is not fantastic. Makes me wonder if we should go back to shelling out from the Rake task to avoid pollution or if there's some other way to segregate things without messing up any other valid use cases